Preparation for 64-bit support
git-svn-id: svn+ssh://svn.code.sf.net/p/jnode/svn/trunk@965 407389f7-7c16-0410-9eee-ad85acb55d7f

+++ core/src/core/org/jnode/vm/x86/MPConfigTable.java
@@ -161,7 +161,7 @@ public class MPConfigTable {
         final int cnt = getEntryCount();
         final ArrayList list = new ArrayList(cnt);
         int offset = 0x2C;
-        final int size = (int)mem.getSize().toInt();
+        final int size = mem.getSize().toInt();
         try {
             while (offset < size) {
                 final int type = mem.getByte(offset);
